Beginnings of the GUI.
I laid out a basic framework for the GUI.
It includes a class called space_object, corresponding to
space-object on the physics end of things.
There's also a class for materials.

A lot of things are currently hard coded (like where all the objects
are, and how many of them there are) which will eventually be
taken as input from the physics engine.