Added a mechanism for the physics engine to accept commands.
It's a function in a let, like this:

(let (stuff)
	(defun stuff ()
		stuff))

so that the function get-command can have a charactar buffer that
persists between calls. I could have made a full on class for this
sort of thing, but I decided not to, for now, for the sake of
simplicity.  I probably will do that though, if I eventually need
to take input from multiple sources.