HUD now aware whether projected point visible, clean up in matrix storage.
The HUD is now aware of whether the projected point is visible,
via a second return value (t for visible, nil for not).

Also, instead of storing a complete projection and modelview matrix
set for each vessel, and projecting (make-vector-3 0 0 0), now a single
projection and modelview matrix set is stored, and (slot-value vessel 'pos)
is plotted instead. This also enables the plotting of arbitrary points,
which is convenient.

I've also included a one-liner sh script, for executing yino.