The Sun is now the source of light, and rough numbers are there for the planets and sun.
Unfortunately, you can't see the sun, because all the light reflecting
off the surface of the sun is only reflecting off the inner surface,
so even though you can see other things by the light of the sun,
the sun itself is invisible.

Some rough numbers from Wikipedia, just to make things more interesting
for testing. This part of the code needs to be cleaned up, and the
numbers updated.