Changed name to TANSTAAFL
TANSTAAFL (pronounced TAN(like the color)-stah(like Bah, Humbug)-fl(like raffle)),
or There Ain't No Such Thing As A Free Lunch, is the new name of the simulator.
It's named partially in honor of Robert Heinlen's great novel "The Moon is a Harsh
Mistress". In this context, TANSTAAFL can refer to the conservation of energy, to
all the work that's gone into making this little bit of free software, or whatever
you think is most appropriate at the time.