Added a material class. Now planets and the Sun are different colors.
I can see the sun! But only if I look directly at it.
There's a problem with the depth interval system.
The near and far viewing limits are specified in terms of straight
line distance from the viewer to the object. However, the near and
far clipping planes are specified in distances straight into the screen.
This should be fixable somehow. I'll work it out.