Adjusted how the thrusters work.
Now you call the generic function (command thruster power-level) to
add power-level to the power level of the thruster, and the generic
function (burn thruster) to burn the thruster at the specified power
level for the timestep.

Thruster groups respond to command and burn in the way you'd expect
(calling said function on all the thrusters in the group),
and calling burn on a vessel will burn all the thruster groups.