Added a bunch of buttons for deleting hud layers
The buttons don't yet have anything but color to show which layer
they go with. Eventually, I'll give the more distinctive labels,
and add buttons for changing the stacking, and the colors.

The bulk of the work in this department is done though. I created
a box-2d subclass that lines up box-2d's sort of like a grid
geometry manager in Tk.