MFDs can now be resized.
Drag the # to move the MFD. Drag the + to resize it. If you resize
the MFD such that the + is to the northwest of the #, it will fail
to create a new viewport, and the MFD will be rendered on the
wrong part of the screen. This should probably be fixed eventually.