Relative hud bug fixed.
It turns out that what I thought was deterministic, intened behavior
in glu-project for computing win-z for points outside the frustrum,
was instead undocumented, random behavior that almost works. This
screwed up the positive and negative labels on the HUD. I fixed it
by doing that part of the computation myself.