Incorporated the button positioning code into the mfd class
There are now 4 lists for subclasses to store buttons in, one list
per side of the MFD. Each side can hold a tunable number of buttons,
currently set to 8. If you want to have some empty space on a side,
you can insert nil into the list for a side, and it will hold an
empty spot.