There is now a button for adding HUD layers
The mechanics are thus:

When a hud is created with make-instance, it should simply appear
if all it's required interesting options are specified. However,
if some options are left blank, and it is not possible to infer
what the user intended with 100% certianty, then an :after
method specialized on initialize-instance should destroy the hud-layer
and instead pop up a menu asking for more input. The selection function
of this menu should then create a new hud-layer with the appropriate
options. This can repeat as many times as required to gather all the
appropriate information. The relative-hud demonstrates this.

There is not yet a mechanism for changing the colors of the newly
created huds. I am not even sure such a mechanism is necessary,
or desirable, though it might be. I need to think it over more.